requirements.txt
# Bangkok PM2.5 Forecasting — Dependencies
# Python 3.12 + ROCm 7.0.2 (AMD Radeon RX 7800 XT / gfx1101)
#
# PyTorch ROCm wheels installed separately from repo.radeon.com:
#   torch==2.8.0+rocm7.0.2
#   torchvision==0.23.0+rocm7.0.2
#   torchaudio==2.8.0+rocm7.0.2
#   triton==3.4.0+rocm7.0.2

# --- Core Data ---
pandas>=3.0.0
numpy>=2.0.0
pyarrow>=23.0.0
requests>=2.32.0

# --- Geospatial ---
geopandas>=1.0.0
shapely>=2.0.0
pyproj>=3.5.0
xarray>=2024.0.0

# --- ML / AI ---
scikit-learn>=1.8.0
shap>=0.50.0

# --- Visualization ---
matplotlib>=3.10.0
seaborn>=0.13.0

# --- Jupyter ---
ipykernel>=7.0.0

# --- Data Version Control ---
dvc>=3.0.0
dvc-s3>=3.0.0

# --- Experiment Tracking ---
mlflow>=2.10.0
optuna>=3.5.0
optuna-integration>=3.5.0

# --- Testing ---
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-benchmark>=4.0.0

# --- Development Tools ---
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
pyyaml>=6.0.0
